test: hold both printers to one equivalence suite over both corpora
`SyntaxBrowserEquivalence.unittest.js` asked an engine whether the
printers' rewrites were equivalent, but only over `configCases`; the wpt
suite asked the same of the web's corpus with its own, smaller
machinery. Fold them together.

The helpers installed into the page, and the comparisons built on them,
move to `test/helpers/syntaxEquivalence.js` — `conditionSignatures`
takes the page and `buildCorpus` takes a directory, so neither is bound
to one corpus. The DOM, `removeImpliedTags` and CSSOM tests are then
parameterized over `configCases` and `wpt` alike, each with its own
filed-defect map; the two IDL tests answer for the tables rather than a
corpus and run once.

CSS is no longer only declarations. A wpt document now goes through the
same `htmlFacets` path, so its inline `<style>` and every `style=""` are
compared as CSSOM — rules, selectors, cascade order, and each at-rule
condition at every size that could tell two apart — rather than as text,
and 460 standalone stylesheets join the 8.8k declarations. The corpus
widens with it: 49,439 documents parsed without an engine, 4,832 handed
to Chrome.

It runs as a spec suite in a job of its own, since it now needs the wpt
submodule; the Deno and Bun runners no longer name it, as neither
matches a `spectest`.
