+++ TESTING_DOCS.md
@@ -79,7 +79,7 @@ This document explains the structure of the `test/` directory in the Webpack pro
 ### 12c. `wpt/`
 
 - **Purpose**: web-platform-tests, read two ways. `html/syntax/parsing/resources/*.dat` is the HTML tree-construction conformance corpus for `parseHtml` (html5lib-tests dropped its copy in `224991e`). The `.html` documents under `html/`, `conformance-checkers/` and `dom/nodes`, plus the declarations the `css/**/parsing/` tests state a verdict for, are the printers' corpus: minifying must not change the DOM webpack's parser builds, the DOM Chrome builds, or the style Chrome computes.
-- **Usage**: Git submodule — initialize with `git submodule update --init --depth 1 test/wpt` (the repository is ~161k files, so keep it shallow). Test runners: `test/html5lib.spectest.js` (`yarn test:html5lib`), which also reads `test/html5lib-tests` — initialize both to run the whole suite — and `test/webPlatformTests.spectest.js` (`yarn test:wpt`), whose browser tier needs a Chrome (`PUPPETEER_EXECUTABLE_PATH` picks a binary other than the installed channel). A document that is not UTF-8 is skipped: the encoding fixtures are UTF-16, which no string API can read as source.
+- **Usage**: Git submodule — initialize with `git submodule update --init --depth 1 test/wpt` (the repository is ~161k files, so keep it shallow). Test runners: `test/html5lib.spectest.js` (`yarn test:html5lib`), which also reads `test/html5lib-tests` — initialize both to run the whole suite — and `test/syntaxEquivalence.spectest.js` (`yarn test:syntax-equivalence`), whose browser tiers need a Chrome (`PUPPETEER_EXECUTABLE_PATH` picks a binary other than the installed channel). A document that is not UTF-8 is skipped: the encoding fixtures are UTF-16, which no string API can read as source.
 
 ### 12d. `css-parsing-tests/`
 
@@ -138,7 +138,7 @@ During the test run, webpack compiles this project and compares the result with 
 
 ### Suites that drive a real browser
 
-`ProfilingPlugin.unittest.js`, `SyntaxBrowserEquivalence.unittest.js` and
+`ProfilingPlugin.unittest.js`, `syntaxEquivalence.spectest.js` and
 `WebpackDevServer.longtest.js` launch Chrome through `test/helpers/launchChrome.js`.
 **A browser that will not launch fails the suite — it is never skipped**, so no
 environment can report these checks as passing without having run them. The
@@ -161,20 +161,20 @@ yarn test
 
 **Choose test command based on modified directory:**
 
-| Modified directory/file   | Command                                                                                                                     |
-| ------------------------- | --------------------------------------------------------------------------------------------------------------------------- |
-| `test/*.unittest.js`      | `yarn test:base --testPathPatterns="<filename>"`                                                                            |
-| `test/cases/`             | `yarn test:basic`                                                                                                           |
-| `test/configCases/`       | `yarn test:basic --testPathPatterns="ConfigTestCases"`                                                                      |
-| `test/statsCases/`        | `yarn test:basic --testPathPatterns="StatsTestCases"`                                                                       |
-| `test/watchCases/`        | `yarn test:base --testPathPatterns="WatchTestCases"`                                                                        |
-| `test/hotCases/`          | `yarn test:base --testPathPatterns="HotTestCases"`                                                                          |
-| `test/benchmarkCases/`    | `FILTER="<case-name>" yarn benchmark`                                                                                       |
-| `lib/runtime/`            | `yarn test:size` (size of the generated code; `--filter "<category>/"` narrows it)                                          |
-| `test/test262-cases/`     | `yarn test:test262` (requires `git submodule update --init test/test262-cases` first)                                       |
-| `test/html5lib-tests/`    | `yarn test:html5lib` (requires `git submodule update --init test/html5lib-tests` first)                                     |
-| `test/wpt/`               | `yarn test:html5lib` + `yarn test:wpt` (require `git submodule update --init --depth 1 test/html5lib-tests test/wpt` first) |
-| `test/css-parsing-tests/` | `yarn test:css-parsing` (requires `git submodule update --init test/css-parsing-tests` first)                               |
+| Modified directory/file   | Command                                                                                                                                    |
+| ------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------ |
+| `test/*.unittest.js`      | `yarn test:base --testPathPatterns="<filename>"`                                                                                           |
+| `test/cases/`             | `yarn test:basic`                                                                                                                          |
+| `test/configCases/`       | `yarn test:basic --testPathPatterns="ConfigTestCases"`                                                                                     |
+| `test/statsCases/`        | `yarn test:basic --testPathPatterns="StatsTestCases"`                                                                                      |
+| `test/watchCases/`        | `yarn test:base --testPathPatterns="WatchTestCases"`                                                                                       |
+| `test/hotCases/`          | `yarn test:base --testPathPatterns="HotTestCases"`                                                                                         |
+| `test/benchmarkCases/`    | `FILTER="<case-name>" yarn benchmark`                                                                                                      |
+| `lib/runtime/`            | `yarn test:size` (size of the generated code; `--filter "<category>/"` narrows it)                                                         |
+| `test/test262-cases/`     | `yarn test:test262` (requires `git submodule update --init test/test262-cases` first)                                                      |
+| `test/html5lib-tests/`    | `yarn test:html5lib` (requires `git submodule update --init test/html5lib-tests` first)                                                    |
+| `test/wpt/`               | `yarn test:html5lib` + `yarn test:syntax-equivalence` (require `git submodule update --init --depth 1 test/html5lib-tests test/wpt` first) |
+| `test/css-parsing-tests/` | `yarn test:css-parsing` (requires `git submodule update --init test/css-parsing-tests` first)                                              |
 
 **Running a single test case** with `--testNamePattern`. The test name format is `<category> <case-name>` (e.g., `css basic`, `asset url`):
