+++ lib/css/syntax.js
@@ -9979,7 +9979,7 @@ const printer = (path, writer) => {
 				if (custom) {
 					// A custom property's value is the text `getPropertyValue()` hands
 					// back, so rewriting it builds a different CSSOM from the same
-					// document — which `SyntaxBrowserEquivalence` forbids, whatever a
+					// document — which `syntaxEquivalence` forbids, whatever a
 					// substitution would read. As written, then; only the comments in it
 					// minify, down to the boundary they stand for.
 					const from = path.start(path.childAt(path.node, 0));
